requirements.txt
fastapi
uvicorn[standard]
transformers>=4.30
accelerate
gradio
diffusers
safetensors
python-multipart
python-dotenv
pydantic
Pillow
regex
